requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pymysql
aiomysql
google-cloud-storage
python-multipart
python-jose
cryptography